# Conflicts:
#	src/main/java/com/wondertek/cpm/service/ContractBudgetService.java

+++ src/main/java/com/wondertek/cpm/web/rest/ContractCostResource.java
@@ -1,7 +1,7 @@
 package com.wondertek.cpm.web.rest;
 
-import java.net.URI;
 import java.net.URISyntaxException;
+import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,7 +18,6 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,9 +25,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.codahale.metrics.annotation.Timed;
+import com.wondertek.cpm.CpmConstants;
+import com.wondertek.cpm.config.StringUtil;
 import com.wondertek.cpm.domain.ContractCost;
 import com.wondertek.cpm.domain.vo.ContractCostVo;
-import com.wondertek.cpm.security.AuthoritiesConstants;
+import com.wondertek.cpm.domain.vo.LongValue;
+import com.wondertek.cpm.security.SecurityUtils;
+import com.wondertek.cpm.service.ContractBudgetService;
 import com.wondertek.cpm.service.ContractCostService;
 import com.wondertek.cpm.web.rest.util.HeaderUtil;
 import com.wondertek.cpm.web.rest.util.PaginationUtil;
@@ -46,50 +49,76 @@ public class ContractCostResource {
         
     @Inject
     private ContractCostService contractCostService;
-
+    @Inject
+    private ContractBudgetService contractBudgetService;
     /**
-     * POST  /contract-costs : Create a new contractCost.
-     *
-     * @param contractCost the contractCost to create
-     * @return the ResponseEntity with status 201 (Created) and with body the new contractCost, or with status 400 (Bad Request) if the contractCost has already an ID
-     * @throws URISyntaxException if the Location URI syntax is incorrect
-     */
-    @PostMapping("/contract-costs")
-    @Timed
-    @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
-    public ResponseEntity<ContractCost> createContractCost(@RequestBody ContractCost contractCost) throws URISyntaxException {
-        log.debug("REST request to save ContractCost : {}", contractCost);
-        if (contractCost.getId() != null) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("contractCost", "idexists", "A new contractCost cannot already have an ID")).body(null);
-        }
-        ContractCost result = contractCostService.save(contractCost);
-        return ResponseEntity.created(new URI("/api/contract-costs/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert("contractCost", result.getId().toString()))
-            .body(result);
-    }
-
-    /**
-     * PUT  /contract-costs : Updates an existing contractCost.
-     *
-     * @param contractCost the contractCost to update
-     * @return the ResponseEntity with status 200 (OK) and with body the updated contractCost,
-     * or with status 400 (Bad Request) if the contractCost is not valid,
-     * or with status 500 (Internal Server Error) if the contractCost couldnt be updated
-     * @throws URISyntaxException if the Location URI syntax is incorrect
+     * 新增和修改
+     * @param contractCost
+     * @return
+     * @throws URISyntaxException
      */
     @PutMapping("/contract-costs")
     @Timed
-    @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
-    public ResponseEntity<ContractCost> updateContractCost(@RequestBody ContractCost contractCost) throws URISyntaxException {
+    public ResponseEntity<Boolean> updateContractCost(@RequestBody ContractCost contractCost) throws URISyntaxException {
         log.debug("REST request to update ContractCost : {}", contractCost);
-        if (contractCost.getId() == null) {
-            return createContractCost(contractCost);
-        }
-        ContractCost result = contractCostService.save(contractCost);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert("contractCost", contractCost.getId().toString()))
-            .body(result);
+        Boolean isNew = contractCost.getId() == null;
+        if (contractCost.getContractId() == null || contractCost.getBudgetId() == null
+        		|| contractCost.getType() == null || contractCost.getCostDay() == null
+        		|| contractCost.getTotal() == null || StringUtil.isNullStr(contractCost.getName())) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.paramNone", "")).body(null);
+		}
+        if (contractCost.getType() == contractCost.TYPE_HUMAN_COST) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.type1Error", "")).body(null);
+		}
+        String updator = SecurityUtils.getCurrentUserLogin();
+        ZonedDateTime updateTime = ZonedDateTime.now();
+        ContractCost newContractCost = contractCostService.getNewContratCost(contractCost);
+        if (isNew) {
+        	newContractCost.setStatus(CpmConstants.STATUS_VALID);
+        	newContractCost.setCreateTime(updateTime);
+        	newContractCost.setCreator(updator);
+		}else {
+			ContractCostVo contractCostVo = contractCostService.getContractCost(newContractCost.getId());
+			if (contractCostVo == null) {
+				return  ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.noPerm", "")).body(null);
+			}
+			ContractCost old = contractCostService.findOne(newContractCost.getId());
+			if (old == null) {
+				return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.idNone", "")).body(null);
+			}else if (old.getContractId() != newContractCost.getContractId().longValue() || old.getBudgetId() != newContractCost.getBudgetId().longValue() ) {
+				return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.contractIdError", "")).body(null);
+			}else if (old.getStatus() == CpmConstants.STATUS_DELETED) {
+				return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.statue2Error", "")).body(null);
+			}else if (old.getType().intValue() != newContractCost.getType()) {
+				return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.type1Error", "")).body(null);
+			}
+			newContractCost.setStatus(old.getStatus());
+			newContractCost.setCreateTime(old.getCreateTime());
+			newContractCost.setCreator(old.getCreator());
+		}
+        newContractCost.setUpdateTime(updateTime);
+        newContractCost.setUpdator(updator);
+        ContractCost result =  contractCostService.save(newContractCost);
+        if (isNew) {
+        	return ResponseEntity.ok()
+                    .headers(HeaderUtil.createEntityCreationAlert("contractCost", result.getId().toString()))
+                    .body(isNew);
+			
+		}else {
+			return ResponseEntity.ok()
+        			.headers(HeaderUtil.createEntityUpdateAlert("contractCost", result.getId().toString()))
+        			.body(isNew);
+		}
     }
+    /**
+     * 页面搜索查询
+     * @param contractId
+     * @param type
+     * @param name
+     * @param pageable
+     * @return
+     * @throws URISyntaxException
+     */
     @GetMapping("/contract-costs")
     @Timed
     @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
@@ -106,11 +135,15 @@ public class ContractCostResource {
         contractCost.setName(name);
         Page<ContractCostVo> page = contractCostService.getUserPage(contractCost,pageable);
         
-        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/project-costs");
+        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/contract-costs");
         
         return new ResponseEntity<List<ContractCostVo>>(page.getContent(), headers,HttpStatus.OK);
     }
-
+    /**
+     * 根据id查询ContractCostVo 返给页面，回显
+     * @param id
+     * @return
+     */
     @GetMapping("/contract-costs/{id}")
     @Timed
     @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
@@ -129,20 +162,30 @@ public class ContractCostResource {
     @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
     public ResponseEntity<Void> deleteContractCost(@PathVariable Long id) {
         log.debug("REST request to delete ContractCost : {}", id);
+        ContractCostVo contractCost = contractCostService.getContractCost(id);
+        if (contractCost != null) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.save.noPerm", "")).body(null);
+		}
+        if (contractCost.getStatus() == CpmConstants.STATUS_DELETED) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.delete.status2Error", "")).body(null);
+		}
+        if (contractCost.getType() == ContractCost.TYPE_HUMAN_COST) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractCost.delete.type1Error", "")).body(null);
+		}
         contractCostService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("contractCost", id.toString())).build();
     }
-
-    @GetMapping("/_search/contract-costs")
-    @Timed
-    @Secured(AuthoritiesConstants.ROLE_CONTRACT_COST)
-    public ResponseEntity<List<ContractCost>> searchContractCosts(@RequestParam String query, @ApiParam Pageable pageable)
-        throws URISyntaxException {
-        log.debug("REST request to search for a page of ContractCosts for query {}", query);
-        Page<ContractCost> page = contractCostService.search(query, pageable);
-        HttpHeaders headers = PaginationUtil.generateSearchPaginationHttpHeaders(query, page, "/api/_search/contract-costs");
-        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
+    
+    @GetMapping("/contract-costs/queryBudges")
+    public ResponseEntity<List<LongValue>> queryBudges(@RequestParam(value = "contractId",required=false) Long contractId){
+    	log.debug("REST request to queryBudges");
+    	List<LongValue> list = contractBudgetService.queryBudges(contractId);
+    	return new ResponseEntity<List<LongValue>>(list, HttpStatus.OK);
     }
-
-
+    @GetMapping("/contract-costs/queryAllBudges")
+    public ResponseEntity<List<LongValue>> queryAllBudges(){
+    	log.debug("REST request to queryBudges");
+    	List<LongValue> list = contractBudgetService.queryBudges();
+    	return new ResponseEntity<List<LongValue>>(list, HttpStatus.OK);
+    }
 }
